fix to reduction rule bug (sf forum id=6033625) and other minor bug fixes

+++ trunk/yawleditor/trunk/source/org/yawlfoundation/yawl/editor/elements/model/YAWLMultipleInstanceTask.java
@@ -25,7 +25,7 @@ package org.yawlfoundation.yawl.editor.elements.model;
 import org.yawlfoundation.yawl.editor.data.DataVariable;
 
 public interface YAWLMultipleInstanceTask {
-  public static final long INFINITY = Long.MAX_VALUE;
+  public static final long INFINITY = Integer.MAX_VALUE;
 
   public static final int STATIC_INSTANCE_CREATION  = -2;
   public static final int DYNAMIC_INSTANCE_CREATION = -3;
